+++ REINS/OpenAddressCircularHash.h
@@ -3,6 +3,7 @@
 #include "SelfMantainedCircularQueue.h"
 #include "Definition.h"
 #include "TreeHash.h"
+#include "QuadTree.h"
 
 class OpenAddressCircularHash : public VirtualHash {
 private:
@@ -11,15 +12,15 @@ private:
 
 public:
 	OpenAddressCircularHash() {}
-	OpenAddressCircularHash(uint _size);
-	void SetupCircularHash(uint _size);
+	OpenAddressCircularHash(unsigned int _size);
+	void SetupCircularHash(unsigned int _size);
 	~OpenAddressCircularHash();
 
 	/*obsolete*/
-	ulong Add(const ulong hashValue, const bool isDuplicated);
+	unsigned long long Add(const unsigned long long hashValue, const bool isDuplicated);
 	/*obsolete*/
-	bool Find(const ulong hashValue);
+	bool Find(const unsigned long long hashValue);
 
-	bool FindAndAdd(const ulong& hashValue, ulong& toBeDel);
+	bool FindAndAdd(const unsigned long long& hashValue, unsigned long long& toBeDel);
 };
