+++ REINS/RedundancyEliminator_CPP.h
@@ -8,13 +8,13 @@ private:
 	CircularHash circHash;
 
 	//Add a new chunk into cache, if hash value queue is full also delete the oldest chunk
-	void addNewChunk(ulong hashValue, char* chunk, uint chunkSize, bool isDuplicate);
-	inline ulong computeChunkHash(char* chunk, uint chunkSize);
+	void addNewChunk(unsigned long long hashValue, char* chunk, unsigned int chunkSize, bool isDuplicate);
+	inline unsigned long long computeChunkHash(char* chunk, unsigned int chunkSize);
 		
 public:
-	deque<uint> chunking(char* package, uint packageSize);
-	uint fingerPrinting(deque<uint> indexQ, char* package);
-	uint eliminateRedundancy(char* package, uint packageSize);
+	deque<unsigned int> chunking(char* package, unsigned int packageSize);
+	unsigned int fingerPrinting(deque<unsigned int> indexQ, char* package);
+	unsigned int eliminateRedundancy(char* package, unsigned int packageSize);
 	RedundancyEliminator_CPP();
 	void SetupRedundancyEliminator_CPP();
 	~RedundancyEliminator_CPP();
